add base.h u8/u16/u32 typedefs for portable unsigned widths (#98)
- centralize tick/seed and byte-flag types in game and world headers
- document LP64 vs DOS width assumptions with compile-time sizeof guards

+++ src/world.c
@@ -239,7 +239,7 @@ void world_init(struct World *world)
     }
 }
 
-void world_step(struct World *world, unsigned long tick)
+void world_step(struct World *world, u32 tick)
 {
     (void)world;
     (void)tick;
